Conflicts:
	.metadata/.log
	.metadata/.mylyn/.taskListIndex/segments_1
	.metadata/.mylyn/.tasks.xml.zip
	.metadata/.mylyn/repositories.xml.zip
	.metadata/.mylyn/tasks.xml.zip
	.metadata/.plugins/org.eclipse.core.resources/.projects/JMineSweeper/.markers
	.metadata/.plugins/org.eclipse.core.resources/.projects/JMineSweeper/org.eclipse.jdt.core/state.dat
	.metadata/.plugins/org.eclipse.core.resources/.projects/LS/.indexes/7/history.index
	.metadata/.plugins/org.eclipse.core.resources/.projects/LS/.location
	.metadata/.plugins/org.eclipse.core.resources/.projects/LS/.markers
	.metadata/.plugins/org.eclipse.core.resources/.projects/LS/org.eclipse.egit.core/GitProjectData.properties
	.metadata/.plugins/org.eclipse.core.resources/.root/.indexes/properties.index
	.metadata/.plugins/org.eclipse.core.resources/.safetable/org.eclipse.core.resources
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.debug.core.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.debug.ui.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.egit.core.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.egit.ui.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.jdt.core.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.jdt.debug.ui.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.jdt.launching.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.jdt.ui.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.recommenders.completion.rcp.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.recommenders.rcp.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.team.ui.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.ui.browser.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.ui.ide.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.ui.workbench.prefs
	.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.wst.sse.ui.prefs
	.metadata/.plugins/org.eclipse.debug.core/.launches/Test1.launch
	.metadata/.plugins/org.eclipse.debug.core/.launches/Test2.launch
	.metadata/.plugins/org.eclipse.debug.core/.launches/Test3.launch
	.metadata/.plugins/org.eclipse.debug.core/.launches/ui.launch
	.metadata/.plugins/org.eclipse.debug.ui/dialog_settings.xml
	.metadata/.plugins/org.eclipse.debug.ui/launchConfigurationHistory.xml
	.metadata/.plugins/org.eclipse.e4.workbench/workbench.xmi
	.metadata/.plugins/org.eclipse.egit.ui/dialog_settings.xml
	.metadata/.plugins/org.eclipse.equinox.p2.ui/dialog_settings.xml
	.metadata/.plugins/org.eclipse.jdt.core/externalFilesCache
	.metadata/.plugins/org.eclipse.jdt.core/externalLibsTimeStamps
	.metadata/.plugins/org.eclipse.jdt.core/nonChainingJarsCache
	.metadata/.plugins/org.eclipse.jdt.core/savedIndexNames.txt
	.metadata/.plugins/org.eclipse.jdt.core/variablesAndContainers.dat
	.metadata/.plugins/org.eclipse.jdt.debug.ui/dialog_settings.xml
	.metadata/.plugins/org.eclipse.jdt.launching/.install.xml
	.metadata/.plugins/org.eclipse.jdt.launching/libraryInfos.xml
	.metadata/.plugins/org.eclipse.jdt.ui/OpenTypeHistory.xml
	.metadata/.plugins/org.eclipse.jdt.ui/QualifiedTypeNameHistory.xml
	.metadata/.plugins/org.eclipse.jdt.ui/dialog_settings.xml
	.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/.workspace/2015/10/42/refactorings.history
	.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/.workspace/2015/10/42/refactorings.index
	.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/.workspace/2015/10/43/refactorings.history
	.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/.workspace/2015/10/43/refactorings.index
	.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/.workspace/2015/10/44/refactorings.history
	.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/.workspace/2015/10/44/refactorings.index
	.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/.workspace/2015/7/31/refactorings.history
	.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/.workspace/2015/7/31/refactorings.index
	.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/LS/2015/10/44/refactorings.history
	.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/LS/2015/10/44/refactorings.index
	.metadata/.plugins/org.eclipse.ltk.ui.refactoring/dialog_settings.xml
	.metadata/.plugins/org.eclipse.m2e.core.ui/dialog_settings.xml
	.metadata/.plugins/org.eclipse.m2e.core/client.lifecyclemapping
	.metadata/.plugins/org.eclipse.m2e.core/nexus/05b0fe8524860bd73cbb07ef30fb34cc/segments.gen
	.metadata/.plugins/org.eclipse.m2e.core/nexus/830bc118332e77292949ed1e6d2fabe0/segments.gen
	.metadata/.plugins/org.eclipse.m2e.core/server.lifecyclemapping
	.metadata/.plugins/org.eclipse.m2e.core/workspaceState.ser
	.metadata/.plugins/org.eclipse.m2e.core/workspacestate.properties
	.metadata/.plugins/org.eclipse.m2e.jdt/client.container
	.metadata/.plugins/org.eclipse.m2e.jdt/server.container
	.metadata/.plugins/org.eclipse.m2e.logback.configuration/0.log
	.metadata/.plugins/org.eclipse.pde.core/.cache/clean-cache.properties
	.metadata/.plugins/org.eclipse.team.ui/dialog_settings.xml
	.metadata/.plugins/org.eclipse.team.ui/syncParticipants.xml
	.metadata/.plugins/org.eclipse.ui.ide/dialog_settings.xml
	.metadata/.plugins/org.eclipse.ui.workbench.texteditor/dialog_settings.xml
	.metadata/.plugins/org.eclipse.ui.workbench/dialog_settings.xml
	.metadata/.plugins/org.eclipse.ui.workbench/workingsets.xml
	.metadata/.plugins/org.eclipse.wst.sse.core/task-tags.properties
	.metadata/.plugins/org.eclipse.wst.xml.core/system_catalog.xml
	.metadata/version.ini
	Logistics-System/bin/bl/managementbl/institutionbl/InstitutionBL.class

+++ Logistics-System/src/PO/CashRepPO.java
@@ -4,18 +4,18 @@ import java.util.ArrayList;
 import VO.*;
 
 public class CashRepPO extends ReceiptPO {
-	private int money;
+	private double money;
 	private String getCourierNum;
 	private ArrayList<GoodsVO> goods;
 	
-	public CashRepPO(String num, String date, int money, String getCourierNum, ArrayList<GoodsVO> goods) {
+	public CashRepPO(String num, String date, double money, String getCourierNum, ArrayList<GoodsVO> goods) {
 		super(num, date);
 		this.money = money;
 		this.getCourierNum = getCourierNum;
 		this.goods = goods;
 	}
 
-	public int getMoney() {
+	public double getMoney() {
 		return money;
 	}
